Add ssl cert and CA support.
Two methods are provided to work with certificates and certificate requests.

A module that calls the openssl binary. This requires managing the config file.

A module that wraps pyOpenSSL to make it easier to use and more consistent with
Python, and Pycopia, style interfaces.

It's not complete, but basically works. Generating CSR with extensions doesn't
seem to work right now with the pyOpenSSL.